requirements.txt
requests
pypdf
python-docx
Pillow
joblib
numpy
scikit-learn
pyobjc
SpeechRecognition
PyAudio
playsound 
beautifulsoup4
requests
trafilatura
beautifulsoup4
lxml
pypdf
python-docx
ollama
sentence-transformers